Modified name of the app
