+++ miniipip/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class MyappConfig(AppConfig):
-    name = 'myapp'
+    name = 'miniipip'
